BE/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
elastic-transport==8.4.0
elasticsearch==8.8.0
frozenlist==1.3.3
idna==3.4
install==1.3.5
JPype1==1.4.1
konlpy==0.6.0
lxml==4.9.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.7
packaging==23.1
pytz==2023.3
regex==2023.5.5
requests==2.31.0
sqlparse==0.4.4
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
yarl==1.9.2
